+++ cpp/gui/mrview/tool/view.cpp
@@ -484,7 +484,7 @@ void View::onVolumeIndexChanged() {
 
   for (int i = 0; i < volume_index_layout->count(); ++i) {
     auto *box = dynamic_cast<SpinBox *>(volume_index_layout->itemAt(i)->widget());
-    box->setValue(image.ndim() > static_cast<size_t>(i + 3) ? image.index(i + 3) : 0);
+    box->setValue(image.ndim() > static_cast<size_t>(i) + 3 ? image.index(static_cast<ssize_t>(i) + 3) : 0);
   }
 }
 
@@ -612,7 +612,7 @@ void View::onSetVolumeIndex() {
 
     for (int i = 0; i < volume_index_layout->count(); ++i) {
       auto *box = dynamic_cast<SpinBox *>(volume_index_layout->itemAt(i)->widget());
-      if (image.ndim() <= static_cast<size_t>(i + 3))
+      if (image.ndim() <= static_cast<size_t>(i) + 3)
         break;
       window().set_image_volume(i + 3, box->value());
     }
